Move the code to a floris namespace within src
