+++ tests/wake_test.py
@@ -11,7 +11,7 @@ CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
 
-from src.wake import Wake
+from src.floris.wake import Wake
 from .sample_inputs import SampleInputs
